backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0

# LLM providers
anthropic>=0.29.0
openai>=1.35.0

# RAG
faiss-cpu>=1.8.0
numpy>=1.26.0

# Python 3.9 compatibility
eval_type_backport>=0.2.0

# Document parsing
pymupdf>=1.24.0

# Optional: only needed if EMBEDDING_PROVIDER=voyage
# voyageai>=0.2.0
